requirements.txt
pandas
langchain
langchain_community
pypdf
langchain-google-genai
python-dotenv
PyPDF2
langchain-google-vertexai
google-generativeai
langchain-astradb
fastapi
uvicorn
jinja2
python-multipart



-e .